[Add] TileBuyUI 스크립트 진행 중
구매 개발 진행 중

+++ Assets/Scripts/ScriptableObjects/TileInfoData.cs
@@ -32,17 +32,17 @@ public class TileInfoData : ScriptableObject
     [Header("지역 가격 색상 타입")] public int _tilePriceColor;
 
     [Separator]
-    [Header("타일 땅 가격")] public int _tileLandPrice;
-    [Header("타일 1번 건물 가격")] public int _tilePensionPrice;
-    [Header("타일 2번 건물 가격")] public int _tileCondoPrice;
-    [Header("타일 3번 건물 가격")] public int _tileHotelPrice;
-    [Header("랜드마크 가격")] public int _tileLandMarkPrice;
+    [Header("타일 땅 가격")] public double _tileLandPrice;
+    [Header("타일 1번 건물 가격")] public double _tilePensionPrice;
+    [Header("타일 2번 건물 가격")] public double _tileCondoPrice;
+    [Header("타일 3번 건물 가격")] public double _tileHotelPrice;
+    [Header("랜드마크 가격")] public double _tileLandMarkPrice;
 
     [Separator]
-    [Header("타일 땅 통행료")] public int _tileLandToll;
-    [Header("타일 1번 건물 통행료")] public int _tilePensionToll;
-    [Header("타일 2번 건물 통행료")] public int _tileCondoToll;
-    [Header("타일 3번 건물 통행료")] public int _tileHotelToll;
-    [Header("랜드마크 통행료")] public int _tileLandMarkToll;
+    [Header("타일 땅 통행료")] public double _tileLandToll;
+    [Header("타일 1번 건물 통행료")] public double _tilePensionToll;
+    [Header("타일 2번 건물 통행료")] public double _tileCondoToll;
+    [Header("타일 3번 건물 통행료")] public double _tileHotelToll;
+    [Header("랜드마크 통행료")] public double _tileLandMarkToll;
 
 }
